Update to support Scenic 0.11
Unfortunately there's no good upgrade guide for the drivers, but many of
the changes are similar to the scene changes:
https://github.com/boydm/scenic/blob/9314020b2962e38bea871e8e1f59cd273dfe0af0/guides/upgrading_to_v0.11.md

I changed the driver to allow Scenic to start the local driver because
it feels a bit cleaner to me.

I've also updated the elixir requirement from 1.8 to 1.11 (which is what
Scenic 0.11 will require)

Unfortunately there's been an update to the format that hex uses to
update mix.lock so the diff there is pretty noisy.